elleandanders_site/requirements/project.txt
Django==1.3
django-form-utils
django-model-utils
-e git+http://github.com/issackelly/django-uni-form.git#egg=django_uni_form